store: move logic from cmd.Store to persistence.Store
The function `cmd.Store` acts as wrapper around `persistence.Store` by
validating & sorting the config before writing it to the file.

To avoid confusion around these two separate functions - this commit
deletes `cmd.Store` in favor of `persistence.Store` and moves the
validation & sorting into `persistence.Store`.
